+++ Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [3.0.0-beta] - 2024-12-21
 
 The package now requires PHP 8.2+ and Postgres 12+. BC breaks are possible due to new language features being used,
 please consult the [upgrade instructions](./UPGRADE.md).
@@ -354,4 +354,4 @@ Initial release on GitHub
 [2.4.0]: https://github.com/sad-spirit/pg-wrapper/compare/v2.3.0...v2.4.0
 [2.4.1]: https://github.com/sad-spirit/pg-wrapper/compare/v2.4.0...v2.4.1
 [2.5.0]: https://github.com/sad-spirit/pg-wrapper/compare/v2.4.1...v2.5.0
-[Unreleased]: https://github.com/sad-spirit/pg-wrapper/compare/v2.5.0...HEAD
+[3.0.0-beta]: https://github.com/sad-spirit/pg-wrapper/compare/v2.5.0...v3.0.0-beta